refactor(doctor): split health contributions
