+++ src/flows/doctor-health-contribution-runners.workspace.ts
@@ -1,0 +1,286 @@
+import type { DoctorOptions } from "../commands/doctor-prompter.js";
+import type { OpenClawConfig } from "../config/types.openclaw.js";
+import type { DoctorHealthFlowContext } from "./doctor-health-contribution-types.js";
+import { resolveDoctorMode } from "./doctor-health-contribution-utils.js";
+import type { HealthCheckContext, HealthFinding } from "./health-checks.js";
+
+type PluginVersionDriftReport =
+  import("../plugins/plugin-version-drift.js").PluginVersionDriftReport;
+
+export async function runHooksModelHealth(ctx: DoctorHealthFlowContext): Promise<void> {
+  if (!ctx.cfg.hooks?.gmail?.model?.trim()) {
+    return;
+  }
+  const { DEFAULT_MODEL, DEFAULT_PROVIDER } = await import("../agents/defaults.js");
+  const { loadPreparedModelCatalog } = await import("../agents/prepared-model-catalog.js");
+  const { getModelRefStatus, resolveConfiguredModelRef, resolveHooksGmailModel } =
+    await import("../agents/model-selection.js");
+  const { note } = await import("../../packages/terminal-core/src/note.js");
+  const hooksModelRef = resolveHooksGmailModel({ cfg: ctx.cfg, defaultProvider: DEFAULT_PROVIDER });
+  if (!hooksModelRef) {
+    note(`- hooks.gmail.model "${ctx.cfg.hooks.gmail.model}" could not be resolved`, "Hooks");
+    return;
+  }
+  const { provider: defaultProvider, model: defaultModel } = resolveConfiguredModelRef({
+    cfg: ctx.cfg,
+    defaultProvider: DEFAULT_PROVIDER,
+    defaultModel: DEFAULT_MODEL,
+  });
+  const catalog = await loadPreparedModelCatalog({ config: ctx.cfg, readOnly: true });
+  const status = getModelRefStatus({
+    cfg: ctx.cfg,
+    catalog,
+    ref: hooksModelRef,
+    defaultProvider,
+    defaultModel,
+  });
+  const warnings: string[] = [];
+  if (!status.allowed) {
+    warnings.push(
+      `- hooks.gmail.model "${status.key}" not allowed by agents.defaults.modelPolicy.allow (will use primary instead)`,
+    );
+  }
+  if (!status.inCatalog) {
+    warnings.push(
+      `- hooks.gmail.model "${status.key}" not in the model catalog (may fail at runtime)`,
+    );
+  }
+  if (warnings.length > 0) {
+    note(warnings.join("\n"), "Hooks");
+  }
+}
+
+export async function runSystemdLingerHealth(ctx: DoctorHealthFlowContext): Promise<void> {
+  if (
+    ctx.options.nonInteractive === true ||
+    process.platform !== "linux" ||
+    resolveDoctorMode(ctx.cfg) !== "local"
+  ) {
+    return;
+  }
+  const { resolveGatewayService } = await import("../daemon/service.js");
+  const { ensureSystemdUserLingerInteractive } = await import("../commands/systemd-linger.js");
+  const { note } = await import("../../packages/terminal-core/src/note.js");
+  let loaded;
+  try {
+    loaded = await resolveGatewayService().isLoaded({ env: process.env });
+  } catch {
+    loaded = false;
+  }
+  if (!loaded) {
+    return;
+  }
+  await ensureSystemdUserLingerInteractive({
+    runtime: ctx.runtime,
+    prompter: { confirm: async (p) => ctx.prompter.confirm(p), note },
+    reason:
+      "Gateway runs as a systemd user service. Without lingering, systemd stops the user session on logout/idle and kills the Gateway.",
+    requireConfirm: true,
+  });
+}
+
+export async function detectSystemdLingerFindings(
+  ctx: HealthCheckContext,
+): Promise<readonly HealthFinding[]> {
+  if (process.platform !== "linux" || resolveDoctorMode(ctx.cfg) !== "local") {
+    return [];
+  }
+  const { resolveGatewayService } = await import("../daemon/service.js");
+  let loaded;
+  try {
+    loaded = await resolveGatewayService().isLoaded({ env: process.env });
+  } catch {
+    loaded = false;
+  }
+  if (!loaded) {
+    return [];
+  }
+  const { isSystemdUserServiceAvailable, readSystemdUserLingerStatus } =
+    await import("../daemon/systemd.js");
+  if (!(await isSystemdUserServiceAvailable(process.env))) {
+    return [];
+  }
+  const status = await readSystemdUserLingerStatus(process.env);
+  if (!status || status.linger === "yes") {
+    return [];
+  }
+  return [
+    {
+      checkId: "core/doctor/systemd-linger",
+      severity: "warning",
+      source: "doctor",
+      message: `Systemd lingering is disabled for ${status.user}.`,
+      target: `systemd.user.${status.user}`,
+      requirement: "systemd user lingering enabled",
+      fixHint: `Run: sudo loginctl enable-linger ${status.user}`,
+    },
+  ];
+}
+
+export async function collectWorkspaceStatusPluginVersionDrift(params: {
+  cfg: OpenClawConfig;
+  options?: Pick<DoctorOptions, "allowExec" | "deep" | "nonInteractive">;
+}): Promise<PluginVersionDriftReport | undefined> {
+  if (params.cfg.gateway?.mode === "remote") {
+    return undefined;
+  }
+  try {
+    const { gatherDaemonStatus } = await import("../cli/daemon-cli/status.gather.js");
+    const status = await gatherDaemonStatus({
+      rpc: { timeout: params.options?.nonInteractive === true ? "3000" : "10000", json: true },
+      probe: true,
+      requireRpc: false,
+      deep: params.options?.deep === true,
+      allowExecSecretRefs: params.options?.allowExec === true,
+    });
+    const hasProbedGatewayVersion =
+      typeof status.gateway?.version === "string" && status.gateway.version.trim() !== "";
+    if (status.pluginVersionDrift && hasProbedGatewayVersion && !status.rpc?.authWarning) {
+      return status.pluginVersionDrift;
+    }
+  } catch {
+    // Best-effort diagnostic: doctor should keep running if daemon status is unavailable.
+  }
+  return undefined;
+}
+
+export async function runWorkspaceStatusHealth(ctx: DoctorHealthFlowContext): Promise<void> {
+  const pluginVersionDrift = await collectWorkspaceStatusPluginVersionDrift({
+    cfg: ctx.cfg,
+    options: ctx.options,
+  });
+  const { noteWorkspaceStatus } = await import("../commands/doctor-workspace-status.js");
+  noteWorkspaceStatus(ctx.cfg, { pluginVersionDrift });
+}
+
+export async function runSkillsHealth(ctx: DoctorHealthFlowContext): Promise<void> {
+  const { maybeRepairSkillReadiness } = await import("../commands/doctor-skills.js");
+  ctx.cfg = await maybeRepairSkillReadiness({ cfg: ctx.cfg, prompter: ctx.prompter });
+}
+
+export async function runBootstrapSizeHealth(ctx: DoctorHealthFlowContext): Promise<void> {
+  const { noteBootstrapFileSize } = await import("../commands/doctor-bootstrap-size.js");
+  await noteBootstrapFileSize(ctx.cfg);
+}
+
+export async function runHeartbeatCadenceMigrationHealth(
+  ctx: DoctorHealthFlowContext,
+): Promise<void> {
+  const { maybeMigrateHeartbeatCadenceToCron } =
+    await import("../commands/doctor-heartbeat-cadence-migration.js");
+  await maybeMigrateHeartbeatCadenceToCron({
+    cfg: ctx.cfg,
+    shouldRepair: ctx.prompter.shouldRepair,
+    env: ctx.env,
+  });
+}
+
+export async function runHeartbeatScratchMigrationHealth(
+  ctx: DoctorHealthFlowContext,
+): Promise<void> {
+  const { maybeMigrateHeartbeatFilesToScratch } =
+    await import("../commands/doctor-heartbeat-scratch-migration.js");
+  await maybeMigrateHeartbeatFilesToScratch({
+    cfg: ctx.cfg,
+    shouldRepair: ctx.prompter.shouldRepair,
+    env: ctx.env,
+  });
+}
+
+export async function runHeartbeatTaskMigrationHealth(ctx: DoctorHealthFlowContext): Promise<void> {
+  const { maybeMigrateHeartbeatTasksToCron } =
+    await import("../commands/doctor-heartbeat-task-migration.js");
+  await maybeMigrateHeartbeatTasksToCron({
+    cfg: ctx.cfg,
+    shouldRepair: ctx.prompter.shouldRepair,
+    env: ctx.env,
+  });
+}
+
+export async function runShellCompletionHealth(ctx: DoctorHealthFlowContext): Promise<void> {
+  const { doctorShellCompletion } = await import("../commands/doctor-completion.js");
+  await doctorShellCompletion(ctx.runtime, ctx.prompter, {
+    nonInteractive: ctx.options.nonInteractive,
+  });
+}
+
+export async function runMemorySearchHealthContribution(
+  ctx: DoctorHealthFlowContext,
+): Promise<void> {
+  const { maybeRepairMemoryRecallHealth, noteMemoryRecallHealth, noteMemorySearchHealth } =
+    await import("../commands/doctor-memory-search.js");
+  if (ctx.prompter.shouldRepair) {
+    await maybeRepairMemoryRecallHealth({ cfg: ctx.cfg, prompter: ctx.prompter });
+  }
+  await noteMemorySearchHealth(ctx.cfg, {
+    gatewayMemoryProbe: ctx.gatewayMemoryProbe ?? { checked: false, ready: false, skipped: false },
+  });
+  if (ctx.options.deep === true) {
+    await noteMemoryRecallHealth(ctx.cfg);
+  }
+}
+
+function memorySearchNoteToFinding(message: string): HealthFinding | null {
+  const lines = message.split("\n");
+  const firstLine = (lines[0] ?? message).trim();
+  if (firstLine === "Memory search is explicitly disabled (enabled: false).") {
+    return null;
+  }
+  const fixHint = lines
+    .slice(1)
+    .map((line) => line.trimEnd())
+    .join("\n")
+    .trim();
+  let path = "memory.search.provider";
+  if (firstLine.includes("No active memory plugin")) {
+    path = "plugins.slots.memory";
+  } else if (firstLine.includes("QMD memory backend")) {
+    path = "memory.backend";
+  } else if (firstLine.includes("OpenAI-compatible embeddings endpoint")) {
+    path = "memory.search.remote.baseUrl";
+  } else if (firstLine.includes("OpenAI-compatible embedding model")) {
+    path = "memory.search.model";
+  }
+  return {
+    checkId: "core/doctor/memory-search",
+    severity: "warning",
+    message: firstLine,
+    path,
+    ...(fixHint ? { fixHint } : {}),
+  };
+}
+
+export async function collectMemorySearchHealthFindings(
+  ctx: HealthCheckContext,
+): Promise<readonly HealthFinding[]> {
+  const { noteMemorySearchHealth } = await import("../commands/doctor-memory-search.js");
+  const notes: string[] = [];
+  await noteMemorySearchHealth(ctx.cfg, {
+    includeWorkspaceMemoryHealth: false,
+    skipQmdBinaryProbe: true,
+    skipAuthProfileResolution: true,
+    gatewayMemoryProbe: { checked: false, ready: false, skipped: true },
+    noteFn: (message) => notes.push(String(message)),
+  });
+  return notes.flatMap((message) => {
+    const finding = memorySearchNoteToFinding(message);
+    return finding ? [finding] : [];
+  });
+}
+
+export async function runWorkspaceSuggestionsHealth(ctx: DoctorHealthFlowContext): Promise<void> {
+  if (ctx.options.workspaceSuggestions === false) {
+    return;
+  }
+  const { resolveAgentWorkspaceDir, resolveDefaultAgentId } =
+    await import("../agents/agent-scope.js");
+  const { noteWorkspaceBackupTip } = await import("../commands/doctor-state-integrity.js");
+  const { MEMORY_SYSTEM_PROMPT, shouldSuggestMemorySystem } =
+    await import("../commands/doctor-workspace.js");
+  const { note } = await import("../../packages/terminal-core/src/note.js");
+  const workspaceDir = resolveAgentWorkspaceDir(ctx.cfg, resolveDefaultAgentId(ctx.cfg));
+  noteWorkspaceBackupTip(workspaceDir);
+  if (await shouldSuggestMemorySystem(workspaceDir)) {
+    note(MEMORY_SYSTEM_PROMPT, "Workspace");
+  }
+}
